+++ src/LaravelAwsSsoServiceProvider.php
@@ -6,13 +6,16 @@ namespace LaravelAwsSso;
 
 use Illuminate\Console\Events\CommandStarting;
 use Illuminate\Contracts\Events\Dispatcher;
+use Illuminate\Foundation\DevCommands;
 use Illuminate\Support\ServiceProvider;
 use LaravelAwsSso\Auth\AwsSsoAuthenticator;
 use LaravelAwsSso\Aws\AwsCli;
 use LaravelAwsSso\Aws\ProcessAwsCli;
 use LaravelAwsSso\Console\LoginCommand;
 use LaravelAwsSso\Console\StatusCommand;
+use LaravelAwsSso\Console\WatchCommand;
 use LaravelAwsSso\Listeners\EnsureAwsSsoAuthentication;
+use LaravelAwsSso\Support\AutomaticAuthentication;
 use LaravelAwsSso\Support\StaticCredentials;
 
 final class LaravelAwsSsoServiceProvider extends ServiceProvider
@@ -38,7 +41,11 @@ final class LaravelAwsSsoServiceProvider extends ServiceProvider
 
         $this->publishes([self::CONFIG => $this->app->configPath('aws-sso.php')], 'aws-sso-config');
 
-        $this->commands([LoginCommand::class, StatusCommand::class]);
+        $this->commands([LoginCommand::class, StatusCommand::class, WatchCommand::class]);
+
+        if ($this->app->make(AutomaticAuthentication::class)->monitorsDevSession()) {
+            DevCommands::artisan('aws-sso:watch', 'aws-sso')->blue();
+        }
 
         $this->app->make(Dispatcher::class)->listen(
             CommandStarting::class,